make changes in post route and added Comment route

+++ routes/postRoute.js
@@ -4,22 +4,24 @@ const Post=require('../models/Post')
 const {check,validationResult}=require('express-validator')
 const authentication=require('../middleware/authentication')
 const upload=require('../middleware/upload')
-router.post('/blog/insert',[
-    check('title',"Blog must have title").not().isEmpty(),
-    check('description',"Blog must some short of description").not().isEmpty(),
-    check('userID',"user id is required").not().isEmpty(),
-    check('categoryID',"It must be one of the Category").not().isEmpty()
-],authentication.verifyUser,upload.single('image'),function(req,res){
+router.post('/blog/insert'
+// ,[
+//     check('title',"Blog must have title").not().isEmpty(),
+//     check('description',"Blog must some short of description").not().isEmpty(),
+//     check('userID',"user id is required").not().isEmpty(),
+//     check('categoryID',"It must be one of the Category").not().isEmpty()
+// ]
+,authentication.verifyUser,upload.single('image'),function(req,res){
 const errors=validationResult(req)
 if(req.file==undefined){
     return res.status(400).json({message:"invalid image Type!!"})
 }
-if(errors.isEmpty()){
+// if(errors.isEmpty()){
     
     const title=req.body.title
     const description=req.body.description
     
-    const userID=req.body.userID
+    const userID=req.user._id
     const categoryID=req.body.categoryID
     
     const data=new Post({
@@ -37,14 +39,18 @@ if(errors.isEmpty()){
         res.status(400).json({message:e})
     })
 
-}
+    // }
+// else{
+//     res.status(400).json(errors.array())
+// }
 })
 router.put('/post/update/:id',authentication.verifyUser,authentication.verifyNormalUser,function(req,res){
     const id=req.params.id;
     const title=req.body.title
     const description=req.body.description
+    const postedBY=req.body.userID
     const image=req.body.image
-
+if(req.user._id==postedBY){
     Post.updateOne({_id:id},{title:title,description:description,image:image})
     .then(function(result){
         res.status(203).json({message:"Updated sucessfully !!"})
@@ -52,6 +58,10 @@ router.put('/post/update/:id',authentication.verifyUser,authentication.verifyNor
     .catch(function(e){
         res.status(404).json({message:e})
     })
+}
+ else{
+     res.status(405).json({message:"your are not allowed to update"})
+ }   
 
 })
 router.get('/post/all',authentication.verifyUser,authentication.verifyAdminNormalUser,
@@ -64,15 +74,22 @@ function(req,res){
         res.status(400).json({message:e})
     })
 })
-router.delete('/post/delete/:id',authentication.verifyUser,authentication.verifyAdminNormalUser,
+router.delete('/post/delete/:id',authentication.verifyUser,
 function(req,res){
     const id=req.params.id
-    Post.deleteOne({_id:id})
-    .then(function(result){
-        res.status(201).json({message:"Deleted Successfully !!"})
-    })
-    .catch(function(e){
-        res.status(500).json({message:e})
-    })
+    const userID=req.body.userID
+    if(req.user._id==userID){
+        Post.deleteOne({_id:id})
+        .then(function(result){
+            res.status(201).json({message:"Deleted Successfully !!"})
+        })
+        .catch(function(e){
+            res.status(500).json({message:e})
+        })
+    }
+    else{
+        res.status(405).json({message:"not allowed to delete the blog"})
+    }
+   
 })
 module.exports=router
